requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Werkzeug==2.2.5
